refactor: remove storage and testing crates
