+++ dpc/src/base_dpc/mod.rs
@@ -76,9 +76,6 @@ pub use transaction::*;
 
 pub mod instantiated;
 
-#[cfg(test)]
-mod tests;
-
 ///////////////////////////////////////////////////////////////////////////////
 
 /// Trait that stores all information about the components of a Plain DPC
