+++ snarkos-integration/Cargo.toml
@@ -1,8 +1,8 @@
 [package]
-name = "snarkvm-dpc"
+name = "snarkvm-snarkos-integration"
 version = "0.2.2"
 authors = [ "The Aleo Team <hello@aleo.org>" ]
-description = "DPC for a decentralized virtual machine"
+description = "A test suite for functionalities requiring snarkOS integration"
 homepage = "https://aleo.org"
 repository = "https://github.com/AleoHQ/snarkVM"
 keywords = [
@@ -18,52 +18,47 @@ license = "GPL-3.0"
 edition = "2018"
 
 [dependencies.snarkvm-algorithms]
-path = "../algorithms"
 version = "0.2.2"
 default-features = false
 
 [dependencies.snarkvm-curves]
-path = "../curves"
+version = "0.2.2"
+default-features = false
+
+[dependencies.snarkvm-dpc]
 version = "0.2.2"
 default-features = false
 
 [dependencies.snarkvm-fields]
-path = "../fields"
 version = "0.2.2"
 default-features = false
 
 [dependencies.snarkvm-gadgets]
-path = "../gadgets"
 version = "0.2.2"
 default-features = false
 
 [dependencies.snarkvm-objects]
-path = "../objects"
 version = "0.2.2"
 default-features = false
 
 [dependencies.snarkvm-parameters]
-path = "../parameters"
 version = "0.2.2"
 default-features = false
 
 [dependencies.snarkvm-profiler]
-path = "../profiler"
 version = "0.2.2"
 default-features = false
 
 [dependencies.snarkvm-r1cs]
-path = "../r1cs"
 version = "0.2.2"
 default-features = false
 
 [dependencies.snarkvm-utilities]
-path = "../utilities"
 version = "0.2.2"
 default-features = false
 
 [dependencies.anyhow]
-version = "1.0.40"
+version = "1.0.38"
 
 [dependencies.base58]
 version = "0.1"
@@ -86,18 +81,19 @@ version = "0.10.0"
 [dependencies.rand]
 version = "0.8"
 
+[dependencies.snarkos-storage]
+git = "https://github.com/ljedrz/snarkOS"
+rev = "2eca11a85989bab2f5ea00854f7efdfab7859d95"
+default-features = false
+features = ["mem_storage"]
+
+[dependencies.snarkos-testing]
+git = "https://github.com/ljedrz/snarkOS"
+rev = "2eca11a85989bab2f5ea00854f7efdfab7859d95"
+default-features = false
+
 [dependencies.thiserror]
 version = "1.0"
 
-[dev-dependencies.rand_xorshift]
+[dependencies.rand_xorshift]
 version = "0.3"
-
-[features]
-default = [
-  "snarkvm-algorithms/default",
-  "snarkvm-curves/default",
-  "snarkvm-gadgets/default",
-  "snarkvm-parameters/default",
-  "snarkvm-utilities/default"
-]
-print-trace = [ "snarkvm-profiler/print-trace" ]
