+++ snarkos-integration/tests/base_dpc.rs
@@ -14,31 +14,34 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::{
-    account::Account,
-    base_dpc::{
-        execute_inner_proof_gadget,
-        execute_outer_proof_gadget,
-        inner_circuit::InnerCircuit,
-        parameters::{NoopProgramSNARKParameters, SystemParameters},
-        program::*,
-        record::record_encryption::*,
-        record_payload::RecordPayload,
-        BaseDPCComponents,
-        TransactionKernel,
-        DPC,
-    },
-    instantiated::*,
-    traits::{AccountScheme, DPCScheme, Program, Record},
-};
+use snarkos_storage::mem::MemDb;
+use snarkos_testing::{dpc::*, storage::*};
 use snarkvm_algorithms::{
     merkle_tree::MerklePath,
     traits::{MerkleParameters, CRH, SNARK},
 };
 use snarkvm_curves::bls12_377::{Fq, Fr};
+use snarkvm_dpc::{
+    account::{Account, AccountViewKey},
+    base_dpc::{
+        execute_inner_proof_gadget,
+        execute_outer_proof_gadget,
+        inner_circuit::InnerCircuit,
+        instantiated::*,
+        parameters::{NoopProgramSNARKParameters, SystemParameters},
+        program::NoopProgram,
+        record::record_encryption::RecordEncryption,
+        record_payload::RecordPayload,
+        BaseDPCComponents,
+        TransactionKernel,
+        DPC,
+    },
+    traits::{AccountScheme, DPCScheme, Program, Record},
+};
 use snarkvm_objects::{
     dpc::DPCTransactions,
-    traits::LedgerScheme,
+    merkle_root,
+    traits::{LedgerScheme, Transaction},
     Block,
     BlockHeader,
     BlockHeaderHash,
@@ -47,7 +50,6 @@ use snarkvm_objects::{
     ProofOfSuccinctWork,
 };
 use snarkvm_r1cs::{ConstraintSystem, TestConstraintSystem};
-use snarkvm_testing::storage::*;
 use snarkvm_utilities::{
     bytes::{FromBytes, ToBytes},
     to_bytes,
@@ -56,8 +58,212 @@ use snarkvm_utilities::{
 use itertools::Itertools;
 use rand::{Rng, SeedableRng};
 use rand_xorshift::XorShiftRng;
+use std::time::{SystemTime, UNIX_EPOCH};
 
-type L = Ledger<Tx, CommitmentMerkleParameters>;
+type L = Ledger<Tx, CommitmentMerkleParameters, MemDb>;
+
+#[test]
+fn base_dpc_integration_test() {
+    let mut rng = XorShiftRng::seed_from_u64(1231275789u64);
+
+    // Generate or load parameters for the ledger, commitment schemes, and CRH
+    let (ledger_parameters, parameters) = setup_or_load_parameters::<_, MemDb>(false, &mut rng);
+
+    // Generate accounts
+    let [genesis_account, recipient, _] = generate_test_accounts::<_, MemDb>(&parameters, &mut rng);
+
+    // Specify network_id
+    let network_id: u8 = 0;
+
+    // Create a genesis block
+
+    let genesis_block = Block {
+        header: BlockHeader {
+            previous_block_hash: BlockHeaderHash([0u8; 32]),
+            merkle_root_hash: MerkleRootHash([0u8; 32]),
+            pedersen_merkle_root_hash: PedersenMerkleRootHash([0u8; 32]),
+            time: 0,
+            difficulty_target: 0x07FF_FFFF_FFFF_FFFF_u64,
+            nonce: 0,
+            proof: ProofOfSuccinctWork::default(),
+        },
+        transactions: DPCTransactions::new(),
+    };
+
+    let ledger = initialize_test_blockchain::<Tx, CommitmentMerkleParameters, MemDb>(ledger_parameters, genesis_block);
+
+    let noop_program_id = to_bytes![
+        ProgramVerificationKeyCRH::hash(
+            &parameters.system_parameters.program_verification_key_crh,
+            &to_bytes![parameters.noop_program_snark_parameters().verification_key].unwrap()
+        )
+        .unwrap()
+    ]
+    .unwrap();
+
+    // Generate dummy input records having as address the genesis address.
+    let old_account_private_keys = vec![genesis_account.private_key.clone(); NUM_INPUT_RECORDS];
+    let mut old_records = vec![];
+    for i in 0..NUM_INPUT_RECORDS {
+        let old_sn_nonce = SerialNumberNonce::hash(
+            &parameters.system_parameters.serial_number_nonce,
+            &[64u8 + (i as u8); 1],
+        )
+        .unwrap();
+        let old_record = DPC::generate_record(
+            &parameters.system_parameters,
+            old_sn_nonce,
+            genesis_account.address.clone(),
+            true, // The input record is dummy
+            0,
+            RecordPayload::default(),
+            noop_program_id.clone(),
+            noop_program_id.clone(),
+            &mut rng,
+        )
+        .unwrap();
+        old_records.push(old_record);
+    }
+
+    // Construct new records.
+
+    // Set the new records' program to be the "always-accept" program.
+    let new_record_owners = vec![recipient.address.clone(); NUM_OUTPUT_RECORDS];
+    let new_is_dummy_flags = vec![false; NUM_OUTPUT_RECORDS];
+    let new_values = vec![10; NUM_OUTPUT_RECORDS];
+    let new_payloads = vec![RecordPayload::default(); NUM_OUTPUT_RECORDS];
+    let new_birth_program_ids = vec![noop_program_id.clone(); NUM_OUTPUT_RECORDS];
+    let new_death_program_ids = vec![noop_program_id.clone(); NUM_OUTPUT_RECORDS];
+
+    let memo = [4u8; 32];
+
+    // Offline execution to generate a DPC transaction kernel
+    let transaction_kernel = <InstantiatedDPC as DPCScheme<L>>::execute_offline(
+        parameters.system_parameters.clone(),
+        old_records,
+        old_account_private_keys,
+        new_record_owners,
+        &new_is_dummy_flags,
+        &new_values,
+        new_payloads,
+        new_birth_program_ids,
+        new_death_program_ids,
+        memo,
+        network_id,
+        &mut rng,
+    )
+    .unwrap();
+
+    let local_data = transaction_kernel.into_local_data();
+
+    // Generate the program proofs
+
+    let noop_program = NoopProgram::<_, <Components as BaseDPCComponents>::NoopProgramSNARK>::new(noop_program_id);
+
+    let mut old_death_program_proofs = vec![];
+    for i in 0..NUM_INPUT_RECORDS {
+        let private_input = noop_program
+            .execute(
+                &parameters.noop_program_snark_parameters.proving_key,
+                &parameters.noop_program_snark_parameters.verification_key,
+                &local_data,
+                i as u8,
+                &mut rng,
+            )
+            .unwrap();
+
+        old_death_program_proofs.push(private_input);
+    }
+
+    let mut new_birth_program_proofs = vec![];
+    for j in 0..NUM_OUTPUT_RECORDS {
+        let private_input = noop_program
+            .execute(
+                &parameters.noop_program_snark_parameters.proving_key,
+                &parameters.noop_program_snark_parameters.verification_key,
+                &local_data,
+                (NUM_INPUT_RECORDS + j) as u8,
+                &mut rng,
+            )
+            .unwrap();
+
+        new_birth_program_proofs.push(private_input);
+    }
+
+    let (new_records, transaction) = InstantiatedDPC::execute_online(
+        &parameters,
+        transaction_kernel,
+        old_death_program_proofs,
+        new_birth_program_proofs,
+        &ledger,
+        &mut rng,
+    )
+    .unwrap();
+
+    // Check that the transaction is serialized and deserialized correctly
+    let transaction_bytes = to_bytes![transaction].unwrap();
+    let recovered_transaction = Tx::read(&transaction_bytes[..]).unwrap();
+
+    assert_eq!(transaction, recovered_transaction);
+
+    {
+        // Check that new_records can be decrypted from the transaction
+
+        let encrypted_records = transaction.encrypted_records();
+        let new_account_private_keys = vec![recipient.private_key; NUM_OUTPUT_RECORDS];
+
+        for ((encrypted_record, private_key), new_record) in
+            encrypted_records.iter().zip(new_account_private_keys).zip(new_records)
+        {
+            let account_view_key = AccountViewKey::from_private_key(
+                &parameters.system_parameters.account_signature,
+                &parameters.system_parameters.account_commitment,
+                &private_key,
+            )
+            .unwrap();
+
+            let decrypted_record =
+                RecordEncryption::decrypt_record(&parameters.system_parameters, &account_view_key, encrypted_record)
+                    .unwrap();
+
+            assert_eq!(decrypted_record, new_record);
+        }
+    }
+
+    // Craft the block
+
+    let previous_block = ledger.get_latest_block().unwrap();
+
+    let mut transactions = DPCTransactions::new();
+    transactions.push(transaction);
+
+    let transaction_ids = transactions.to_transaction_ids().unwrap();
+
+    let mut merkle_root_bytes = [0u8; 32];
+    merkle_root_bytes[..].copy_from_slice(&merkle_root(&transaction_ids));
+
+    let time = SystemTime::now()
+        .duration_since(UNIX_EPOCH)
+        .expect("Time went backwards")
+        .as_secs() as i64;
+
+    let header = BlockHeader {
+        previous_block_hash: previous_block.header.get_hash(),
+        merkle_root_hash: MerkleRootHash(merkle_root_bytes),
+        time,
+        difficulty_target: previous_block.header.difficulty_target,
+        nonce: 0,
+        pedersen_merkle_root_hash: PedersenMerkleRootHash([0u8; 32]),
+        proof: ProofOfSuccinctWork::default(),
+    };
+
+    assert!(InstantiatedDPC::verify_transactions(&parameters, &transactions.0, &ledger).unwrap());
+
+    let block = Block { header, transactions };
+
+    ledger.insert_and_commit(&block).unwrap();
+    assert_eq!(ledger.len(), 2);
+}
 
 /// Generates and returns noop program parameters and its corresponding program id.
 fn generate_test_noop_program_parameters<R: Rng>(
@@ -195,7 +401,7 @@ fn test_execute_base_dpc_constraints() {
     };
 
     // Use genesis record, serial number, and memo to initialize the ledger.
-    let ledger = initialize_test_blockchain::<Tx, CommitmentMerkleParameters>(ledger_parameters, genesis_block);
+    let ledger = initialize_test_blockchain::<Tx, CommitmentMerkleParameters, MemDb>(ledger_parameters, genesis_block);
 
     let sn_nonce = SerialNumberNonce::hash(&system_parameters.serial_number_nonce, &[0u8; 1]).unwrap();
     let old_record = DPC::generate_record(
@@ -402,7 +608,7 @@ fn test_execute_base_dpc_constraints() {
     let inner_snark_vk: <<Components as BaseDPCComponents>::InnerSNARK as SNARK>::VerificationParameters =
         inner_snark_parameters.1.clone().into();
 
-    let inner_circuit_id = InnerSNARKVerificationKeyCRH::hash(
+    let inner_snark_id = InnerSNARKVerificationKeyCRH::hash(
         &system_parameters.inner_snark_verification_key_crh,
         &to_bytes![inner_snark_vk].unwrap(),
     )
@@ -457,7 +663,7 @@ fn test_execute_base_dpc_constraints() {
         &program_commitment,
         &program_randomness,
         &local_data_root,
-        &inner_circuit_id,
+        &inner_snark_id,
     )
     .unwrap();
 
@@ -478,6 +684,4 @@ fn test_execute_base_dpc_constraints() {
     println!("=========================================================");
 
     assert!(pf_check_cs.is_satisfied());
-
-    kill_storage(ledger);
 }
