--- storage/Cargo.toml
@@ -1,69 +1,0 @@
-[package]
-name = "snarkvm-storage"
-version = "0.2.2"
-authors = [ "The Aleo Team <hello@aleo.org>" ]
-description = "Storage for a decentralized virtual machine"
-homepage = "https://aleo.org"
-repository = "https://github.com/AleoHQ/snarkVM"
-keywords = [
-  "aleo",
-  "cryptography",
-  "blockchain",
-  "decentralized",
-  "zero-knowledge"
-]
-categories = [ "cryptography::cryptocurrencies", "operating-systems" ]
-include = [ "Cargo.toml", "src", "README.md", "LICENSE.md" ]
-license = "GPL-3.0"
-edition = "2018"
-
-[dependencies.snarkvm-algorithms]
-path = "../algorithms"
-version = "0.2.2"
-
-[dependencies.snarkvm-dpc]
-path = "../dpc"
-version = "0.2.2"
-
-[dependencies.snarkvm-objects]
-path = "../objects"
-version = "0.2.2"
-
-[dependencies.snarkvm-parameters]
-path = "../parameters"
-version = "0.2.2"
-
-[dependencies.snarkvm-utilities]
-path = "../utilities"
-version = "0.2.2"
-
-[dependencies.anyhow]
-version = "1.0.40"
-
-[dependencies.bincode]
-version = "1.3.2"
-
-[dependencies.hex]
-version = "0.4.3"
-
-[dependencies.parking_lot]
-version = "0.11.1"
-
-[dependencies.rand]
-version = "0.8"
-
-[dependencies.rocksdb]
-version = "0.15.0"
-
-[dependencies.serde]
-version = "1.0"
-features = [ "derive" ]
-
-[dependencies.thiserror]
-version = "1.0"
-
-[dev-dependencies.snarkvm-curves]
-path = "../curves"
-
-[dev-dependencies.snarkvm-testing]
-path = "../testing"
