--- storage/README.md
@@ -1,5 +1,0 @@
-# snarkvm-storage
-
-[![Crates.io](https://img.shields.io/crates/v/snarkvm-storage.svg?color=neon)](https://crates.io/crates/snarkvm-storage)
-[![Authors](https://img.shields.io/badge/authors-Aleo-orange.svg)](../AUTHORS)
-[![License](https://img.shields.io/badge/License-GPLv3-blue.svg)](./LICENSE.md)
