--- storage/src/objects/mod.rs
@@ -1,43 +1,0 @@
-// Copyright (C) 2019-2021 Aleo Systems Inc.
-// This file is part of the snarkVM library.
-
-// The snarkVM library is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
-// The snarkVM library is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU General Public License for more details.
-
-// You should have received a copy of the GNU General Public License
-// along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
-
-/// The methods defined in this module require direct access to the storage module.
-pub mod block;
-pub use block::*;
-
-pub mod block_header;
-pub use block_header::*;
-
-pub mod block_path;
-pub use block_path::*;
-
-pub mod dpc_state;
-pub use dpc_state::*;
-
-pub mod insert_commit;
-pub use insert_commit::*;
-
-pub mod ledger_scheme;
-pub use ledger_scheme::*;
-
-pub mod memory_pool;
-pub use memory_pool::*;
-
-pub mod records;
-pub use records::*;
-
-pub mod transaction;
-pub use transaction::*;
