--- testing/Cargo.toml
@@ -1,55 +1,0 @@
-[package]
-name = "snarkvm-testing"
-version = "0.2.2"
-authors = [ "The Aleo Team <hello@aleo.org>" ]
-description = "Testing suite for a decentralized virtual machine"
-homepage = "https://aleo.org"
-repository = "https://github.com/AleoHQ/snarkVM"
-keywords = [
-  "aleo",
-  "cryptography",
-  "blockchain",
-  "decentralized",
-  "zero-knowledge"
-]
-categories = [ "cryptography::cryptocurrencies", "operating-systems" ]
-include = [ "Cargo.toml", "src", "README.md", "LICENSE.md" ]
-license = "GPL-3.0"
-edition = "2018"
-
-[dependencies.snarkvm-algorithms]
-path = "../algorithms"
-version = "0.2.2"
-
-[dependencies.snarkvm-curves]
-path = "../curves"
-version = "0.2.2"
-
-[dependencies.snarkvm-dpc]
-path = "../dpc"
-version = "0.2.2"
-
-[dependencies.snarkvm-objects]
-path = "../objects"
-version = "0.2.2"
-
-[dependencies.snarkvm-parameters]
-path = "../parameters"
-version = "0.2.2"
-
-[dependencies.snarkvm-utilities]
-path = "../utilities"
-version = "0.2.2"
-
-[dependencies.snarkvm-storage]
-path = "../storage"
-version = "0.2.2"
-
-[dependencies.once_cell]
-version = "1.7.2"
-
-[dependencies.rand]
-version = "0.8"
-
-[dependencies.rand_xorshift]
-version = "0.3"
