--- testing/README.md
@@ -1,5 +1,0 @@
-# snarkvm-testing
-
-[![Crates.io](https://img.shields.io/crates/v/snarkvm-testing.svg?color=neon)](https://crates.io/crates/snarkvm-testing)
-[![Authors](https://img.shields.io/badge/authors-Aleo-orange.svg)](../AUTHORS)
-[![License](https://img.shields.io/badge/License-GPLv3-blue.svg)](./LICENSE.md)
