--- testing/src/storage/mod.rs
@@ -1,76 +1,0 @@
-// Copyright (C) 2019-2021 Aleo Systems Inc.
-// This file is part of the snarkVM library.
-
-// The snarkVM library is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
-// The snarkVM library is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU General Public License for more details.
-
-// You should have received a copy of the GNU General Public License
-// along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
-
-use snarkvm_algorithms::traits::merkle_tree::LoadableMerkleParameters;
-use snarkvm_dpc::base_dpc::instantiated::{CommitmentMerkleParameters, Tx};
-use snarkvm_objects::{
-    traits::{LedgerScheme, Transaction},
-    Block,
-};
-pub use snarkvm_storage::Ledger;
-
-use rand::{thread_rng, Rng};
-use std::{path::PathBuf, sync::Arc};
-
-pub type Store = Ledger<Tx, CommitmentMerkleParameters>;
-
-pub fn random_storage_path() -> String {
-    let random_path: usize = thread_rng().gen();
-    format!("./test_db-{}", random_path)
-}
-
-// Initialize a test blockchain given genesis attributes
-pub fn initialize_test_blockchain<T: Transaction, P: LoadableMerkleParameters>(
-    parameters: P,
-    genesis_block: Block<T>,
-) -> Ledger<T, P> {
-    let mut path = std::env::temp_dir();
-    path.push(random_storage_path());
-
-    Ledger::<T, P>::destroy_storage(path.clone()).unwrap();
-
-    Ledger::<T, P>::new(Some(&path), parameters, genesis_block).unwrap()
-}
-
-// Open a test blockchain from stored genesis attributes
-pub fn open_test_blockchain<T: Transaction, P: LoadableMerkleParameters>() -> (Arc<Ledger<T, P>>, PathBuf) {
-    let mut path = std::env::temp_dir();
-    path.push(random_storage_path());
-
-    Ledger::<T, P>::destroy_storage(path.clone()).unwrap();
-
-    let storage = Arc::new(Ledger::<T, P>::open_at_path(path.clone()).unwrap());
-
-    (storage, path)
-}
-
-pub fn kill_storage<T: Transaction, P: LoadableMerkleParameters>(ledger: Ledger<T, P>) {
-    let path = ledger.storage.db.path().to_owned();
-
-    drop(ledger);
-    Ledger::<T, P>::destroy_storage(path).unwrap();
-}
-
-pub fn kill_storage_async<T: Transaction, P: LoadableMerkleParameters>(path: PathBuf) {
-    Ledger::<T, P>::destroy_storage(path).unwrap();
-}
-
-pub fn kill_storage_sync<T: Transaction, P: LoadableMerkleParameters>(ledger: Arc<Ledger<T, P>>) {
-    let path = ledger.storage.db.path().to_owned();
-
-    drop(ledger);
-    Ledger::<T, P>::destroy_storage(path).unwrap();
-}
